+++ R/emptyforest.R
@@ -4,6 +4,16 @@
 #' 
 #' @param ntree,nshrub Number of tree and shrub cohorts, respectively.
 #' @param nseed Number of species in the seed bank.
+#' @param addcolumns A character vector with additional columns. Currently allowed are:
+#' \itemize{
+#'  \item{\code{Z100}: A numeric vector with maximum rooting depth in mm.}
+#'  \item{\code{LAI}: Leaf area index (m2/m2).}
+#'  \item{\code{FoliarBiomass}: Standing dry biomass of leaves (kg/m2).}
+#'  \item{\code{FuelLoading}: Fine fuel loading (kg/m2).}
+#'  \item{\code{CrownRatio}: The ratio between crown length and total height (between 0 and 1)}
+#'  \item{\code{Age}: A numeric vector indicating age of cohorts in years.}
+#'  \item{\code{ObsID} A character vector to label specific cohorts in simulations of forest dynamics.}
+#' }
 #' 
 #' @return An empty \code{\link{forest}} object.
 #' 
@@ -16,8 +26,12 @@
 #' # Initializes forest with 2 tree cohorts and 1 shrub cohort
 #' emptyforest(ntree = 2, nshrub = 1)
 #' 
+#' # Initializes with extra column for leaf area index (LAI)
+#' emptyforest(ntree = 2, nshrub = 1, addcolumns = "LAI")
+#' 
 #' @name emptyforest
-emptyforest <- function(ntree = 0, nshrub = 0, nseed = 0) {
+emptyforest <- function(ntree = 0, nshrub = 0, nseed = 0,
+                        addcolumns = NULL) {
   l <- list()
   l$treeData <- data.frame(Species=as.character(rep(NA, ntree)),
                            DBH=as.numeric(rep(NA, ntree)), 
@@ -30,8 +44,40 @@ emptyforest <- function(ntree = 0, nshrub = 0, nseed = 0) {
                             Cover = as.numeric(rep(NA, nshrub)), 
                             Z50 = as.numeric(rep(NA, nshrub)), 
                             Z95=as.numeric(rep(NA, nshrub)))
-  l$herbCover <- NA;
-  l$herbHeight <- NA;
+  if(!is.null(addcolumns)) {
+    if(!is.character(addcolumns)) stop("`addcolumns` should be a character vector")
+    addcolumns <- match.arg(addcolumns, c("Z100", "LAI", "FoliarBiomass", "FuelLoading", "CrownRatio", "Age", "ObsID") , several.ok = TRUE)
+    if("Z100" %in% addcolumns) {
+      l$treeData$Z100 <- as.numeric(rep(NA, ntree))
+      l$shrubData$Z100 <- as.numeric(rep(NA, nshrub))
+    }
+    if("LAI" %in% addcolumns) {
+      l$treeData$LAI <- as.numeric(rep(NA, ntree))
+      l$shrubData$LAI <- as.numeric(rep(NA, nshrub))
+    }
+    if("FoliarBiomass" %in% addcolumns) {
+      l$treeData$FoliarBiomass <- as.numeric(rep(NA, ntree))
+      l$shrubData$FoliarBiomass <- as.numeric(rep(NA, nshrub))
+    }
+    if("FuelLoading" %in% addcolumns) {
+      l$treeData$FuelLoading <- as.numeric(rep(NA, ntree))
+      l$shrubData$FuelLoading <- as.numeric(rep(NA, nshrub))
+    }
+    if("CrownRatio" %in% addcolumns) {
+      l$treeData$CrownRatio <- as.numeric(rep(NA, ntree))
+      l$shrubData$CrownRatio <- as.numeric(rep(NA, nshrub))
+    }
+    if("Age" %in% addcolumns) {
+      l$treeData$Age <- as.numeric(rep(NA, ntree))
+      l$shrubData$Age <- as.numeric(rep(NA, nshrub))
+    }
+    if("ObsID" %in% addcolumns) {
+      l$treeData$ObsID <- as.character(rep(NA, ntree))
+      l$shrubData$ObsID <- as.character(rep(NA, nshrub))
+    }
+  }
+  l$herbCover <- NA
+  l$herbHeight <- NA
   l$seedBank <- data.frame(Species = as.character(rep(NA, nseed)),
                            Percent = as.numeric(rep(NA, nseed)))
   class(l)<-c("forest","list")