+++ R/fordynutils.R
@@ -1,7 +1,7 @@
 .nextYearForest<-function(forest, xo, SpParams, control,
-                          planted_forest = emptyforest(),
-                          recr_forest = emptyforest(),
-                          resp_forest = emptyforest()) {
+                          planted_forest = emptyforest(addcolumns = c("Z100", "Age", "ObsID")),
+                          recr_forest = emptyforest(addcolumns = c("Z100", "Age", "ObsID")),
+                          resp_forest = emptyforest(addcolumns = c("Z100", "Age", "ObsID"))) {
   
   #Determine offsets
   treeOffset <- 0
@@ -22,15 +22,15 @@
   # print(forest)
   if(control$removeEmptyCohorts) {
     emptyTrees <- (forest$treeData$N < control$minimumTreeCohortDensity)
-    if(control$keepCohortsWithObsID) { # Exclude from removal if cohort has non-missing ID
-      if("ID" %in% names(forest$treeData)) {
+    if(control$keepCohortsWithObsID) { # Exclude from removal if cohort has non-missing ObsID
+      if("ObsID" %in% names(forest$treeData)) {
         emptyTrees[!is.na(forest$treeData)] <- FALSE
       }
     }
     if(control$shrubDynamics) {
       emptyShrubs <- (forest$shrubData$Cover < control$minimumShrubCohortCover)
-      if(control$keepCohortsWithObsID) { # Exclude from removal if cohort has non-missing ID
-        if("ID" %in% names(forest$shrubData)) {
+      if(control$keepCohortsWithObsID) { # Exclude from removal if cohort has non-missing ObsID
+        if("ObsID" %in% names(forest$shrubData)) {
           emptyShrubs[!is.na(forest$shrubData)] <- FALSE
         }
       }
@@ -120,55 +120,19 @@
   sel_vec <- c(rep(TRUE, nrow(forest$treeData)),
                rep(control$shrubDynamics, nrow(forest$shrubData)))
   
-  # 4. Merge cohorts in forest object
-  if("ObsID" %in% names(forest$treeData)) {
-    planted_forest$treeData$ObsID <- rep(as.character(NA), nrow(planted_forest$treeData))
-    recr_forest$treeData$ObsID <- rep(as.character(NA), nrow(recr_forest$treeData))
-    resp_forest$treeData$ObsID <- rep(as.character(NA), nrow(resp_forest$treeData))
+  # 4.1 Increase age of current forest 
+  if("Age" %in% names(forest$treeData)) {
+    forest$treeData$Age <- forest$treeData$Age + 1
   }
-  if("ObsID" %in% names(forest$shrubData)) {
-    planted_forest$shrubData$ObsID <- rep(as.character(NA), nrow(planted_forest$shrubData))
-    recr_forest$shrubData$ObsID <- rep(as.character(NA), nrow(recr_forest$shrubData))
-    resp_forest$shrubData$ObsID <- rep(as.character(NA), nrow(resp_forest$shrubData))
+  if("Age" %in% names(forest$shrubData)) {
+    forest$shrubData$Age <- forest$shrubData$Age + 1
   }
   
-  # Age is only managed when cohorts are not dynamically merged
-  if(!control$dynamicallyMergeCohorts) {
-    ##Increase age of current forest
-    if(!("Age" %in% names(forest$treeData))) {
-      forest$treeData$Age <- as.numeric(rep(NA, nrow(forest$treeData)))
-    } else {
-      forest$treeData$Age <- forest$treeData$Age + 1
-    }
-    if(!("Age" %in% names(forest$shrubData))) {
-      forest$shrubData$Age <- as.numeric(rep(NA, nrow(forest$shrubData)))
-    } else {
-      forest$shrubData$Age <- forest$shrubData$Age + 1
-    }
-    ## Define age if missing in planted, resprouts or recruits
-    if(!("Age" %in% names(planted_forest$treeData))) {
-      planted_forest$treeData$Age <- as.numeric(rep(NA, nrow(planted_forest$treeData)))
-    }
-    if(!("Age" %in% names(planted_forest$shrubData))) {
-      planted_forest$shrubData$Age <- as.numeric(rep(NA, nrow(planted_forest$shrubData)))
-    }
-    if(!("Age" %in% names(recr_forest$treeData))) {
-      recr_forest$treeData$Age <- as.numeric(rep(NA, nrow(recr_forest$treeData)))
-    }
-    if(!("Age" %in% names(recr_forest$shrubData))) {
-      recr_forest$shrubData$Age <- as.numeric(rep(NA, nrow(recr_forest$shrubData)))
-    }
-    if(!("Age" %in% names(resp_forest$treeData))) {
-      resp_forest$treeData$Age <- as.numeric(rep(NA, nrow(resp_forest$treeData)))
-    }
-    if(!("Age" %in% names(resp_forest$shrubData))) {
-      resp_forest$shrubData$Age <- as.numeric(rep(NA, nrow(resp_forest$shrubData)))
-    }
-  }
-  
-  ## Merge forests
+  # 4.2 Merge cohorts in forest object
   forest$treeData <- rbind(forest$treeData, planted_forest$treeData, recr_forest$treeData, resp_forest$treeData)
+  row.names(forest$treeData) <- NULL
   forest$shrubData <- rbind(forest$shrubData, planted_forest$shrubData, recr_forest$shrubData, resp_forest$shrubData)
+  row.names(forest$shrubData) <- NULL
   
   # 5.1 Prepare growth input for next year
   FCCSprops = fuel_FCCS(forest, SpParams);